29 - Kullanıcı Yetkileri

+++ post/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -34,6 +34,10 @@ def post_create(request):
     #     # Formu kullanıcıya göster
     #     form = PostForm()
 
+    if not request.user.is_authenticated():
+        # Eğer kullanıcı giriş yapmamış ise hata sayfası gönder
+        return Http404()
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         post = form.save()
@@ -48,6 +52,11 @@ def post_create(request):
 
 
 def post_update(request, id):
+
+    if not request.user.is_authenticated():
+        # Eğer kullanıcı giriş yapmamış ise hata sayfası gönder
+        return Http404()
+
     post = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid():
@@ -63,6 +72,11 @@ def post_update(request, id):
 
 
 def post_delete(request, id):
+
+    if not request.user.is_authenticated():
+        # Eğer kullanıcı giriş yapmamış ise hata sayfası gönder
+        return Http404()
+
     post = get_object_or_404(Post, id=id)
     post.delete()
     return redirect("post:index")
